requirements.txt
openai
python-dotenv
requests
llama-index
llama-index-llms-anthropic
llama-index-embeddings-huggingface
